football-frenzy/requirements.txt
beautifulsoup4==4.13.3
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cs50==9.4.0
Flask==3.1.0
Flask-Session==0.8.0
googlesearch-python==1.3.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
termcolor==2.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
